refactor: remove duplication with constructor and exporter container menu
