+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/containermenu/ConstructorContainerMenu.java
@@ -1,6 +1,5 @@
 package com.refinedmods.refinedstorage2.platform.common.containermenu;
 
-import com.refinedmods.refinedstorage2.platform.common.block.entity.SchedulingModeType;
 import com.refinedmods.refinedstorage2.platform.common.block.entity.UpgradeContainer;
 import com.refinedmods.refinedstorage2.platform.common.block.entity.constructor.ConstructorBlockEntity;
 import com.refinedmods.refinedstorage2.platform.common.containermenu.property.ClientProperty;
@@ -9,7 +8,6 @@ import com.refinedmods.refinedstorage2.platform.common.containermenu.property.Se
 import com.refinedmods.refinedstorage2.platform.common.content.Menus;
 import com.refinedmods.refinedstorage2.platform.common.internal.resource.filter.ResourceFilterContainer;
 import com.refinedmods.refinedstorage2.platform.common.internal.upgrade.UpgradeDestinations;
-import com.refinedmods.refinedstorage2.platform.common.util.RedstoneMode;
 
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.network.chat.MutableComponent;
@@ -18,7 +16,7 @@ import net.minecraft.world.entity.player.Player;
 
 import static com.refinedmods.refinedstorage2.platform.common.util.IdentifierUtil.createTranslation;
 
-public class ConstructorContainerMenu extends AbstractSimpleFilterContainerMenu<ConstructorBlockEntity> {
+public class ConstructorContainerMenu extends AbstractSchedulingContainerMenu<ConstructorBlockEntity> {
     private static final MutableComponent FILTER_HELP = createTranslation("gui", "constructor.filter_help");
 
     public ConstructorContainerMenu(final int syncId, final Inventory playerInventory, final FriendlyByteBuf buf) {
@@ -50,29 +48,13 @@ public class ConstructorContainerMenu extends AbstractSimpleFilterContainerMenu<
 
     @Override
     protected void registerClientProperties() {
-        registerProperty(new ClientProperty<>(PropertyTypes.FUZZY_MODE, false));
-        registerProperty(new ClientProperty<>(PropertyTypes.REDSTONE_MODE, RedstoneMode.IGNORE));
-        registerProperty(new ClientProperty<>(PropertyTypes.SCHEDULING_MODE, SchedulingModeType.DEFAULT));
+        super.registerClientProperties();
         registerProperty(new ClientProperty<>(PropertyTypes.CONSTRUCTOR_DROP_ITEMS, false));
     }
 
     @Override
     protected void registerServerProperties(final ConstructorBlockEntity blockEntity) {
-        registerProperty(new ServerProperty<>(
-            PropertyTypes.FUZZY_MODE,
-            blockEntity::isFuzzyMode,
-            blockEntity::setFuzzyMode
-        ));
-        registerProperty(new ServerProperty<>(
-            PropertyTypes.REDSTONE_MODE,
-            blockEntity::getRedstoneMode,
-            blockEntity::setRedstoneMode
-        ));
-        registerProperty(new ServerProperty<>(
-            PropertyTypes.SCHEDULING_MODE,
-            blockEntity::getSchedulingModeType,
-            blockEntity::setSchedulingModeType
-        ));
+        super.registerServerProperties(blockEntity);
         registerProperty(new ServerProperty<>(
             PropertyTypes.CONSTRUCTOR_DROP_ITEMS,
             blockEntity::isDropItems,
