+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/containermenu/ExporterContainerMenu.java
@@ -1,15 +1,10 @@
 package com.refinedmods.refinedstorage2.platform.common.containermenu;
 
-import com.refinedmods.refinedstorage2.platform.common.block.entity.SchedulingModeType;
 import com.refinedmods.refinedstorage2.platform.common.block.entity.UpgradeContainer;
 import com.refinedmods.refinedstorage2.platform.common.block.entity.exporter.ExporterBlockEntity;
-import com.refinedmods.refinedstorage2.platform.common.containermenu.property.ClientProperty;
-import com.refinedmods.refinedstorage2.platform.common.containermenu.property.PropertyTypes;
-import com.refinedmods.refinedstorage2.platform.common.containermenu.property.ServerProperty;
 import com.refinedmods.refinedstorage2.platform.common.content.Menus;
 import com.refinedmods.refinedstorage2.platform.common.internal.resource.filter.ResourceFilterContainer;
 import com.refinedmods.refinedstorage2.platform.common.internal.upgrade.UpgradeDestinations;
-import com.refinedmods.refinedstorage2.platform.common.util.RedstoneMode;
 
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.network.chat.MutableComponent;
@@ -18,8 +13,7 @@ import net.minecraft.world.entity.player.Player;
 
 import static com.refinedmods.refinedstorage2.platform.common.util.IdentifierUtil.createTranslation;
 
-// TODO: Make more generic?! it's duplicated atm.
-public class ExporterContainerMenu extends AbstractSimpleFilterContainerMenu<ExporterBlockEntity> {
+public class ExporterContainerMenu extends AbstractSchedulingContainerMenu<ExporterBlockEntity> {
     private static final MutableComponent FILTER_HELP = createTranslation("gui", "exporter.filter_help");
 
     public ExporterContainerMenu(final int syncId, final Inventory playerInventory, final FriendlyByteBuf buf) {
@@ -47,31 +41,5 @@ public class ExporterContainerMenu extends AbstractSimpleFilterContainerMenu<Exp
             exporter,
             FILTER_HELP
         );
-    }
-
-    @Override
-    protected void registerClientProperties() {
-        registerProperty(new ClientProperty<>(PropertyTypes.FUZZY_MODE, false));
-        registerProperty(new ClientProperty<>(PropertyTypes.REDSTONE_MODE, RedstoneMode.IGNORE));
-        registerProperty(new ClientProperty<>(PropertyTypes.SCHEDULING_MODE, SchedulingModeType.DEFAULT));
-    }
-
-    @Override
-    protected void registerServerProperties(final ExporterBlockEntity blockEntity) {
-        registerProperty(new ServerProperty<>(
-            PropertyTypes.FUZZY_MODE,
-            blockEntity::isFuzzyMode,
-            blockEntity::setFuzzyMode
-        ));
-        registerProperty(new ServerProperty<>(
-            PropertyTypes.REDSTONE_MODE,
-            blockEntity::getRedstoneMode,
-            blockEntity::setRedstoneMode
-        ));
-        registerProperty(new ServerProperty<>(
-            PropertyTypes.SCHEDULING_MODE,
-            blockEntity::getSchedulingModeType,
-            blockEntity::setSchedulingModeType
-        ));
     }
 }
